fix: background local test also via channel
current implementation for `standalone` setups
was blocking the `perf drive`.

Bonus: remove all old unused complicated code.
